+++ CHANGELOG.md
@@ -5,15 +5,18 @@
 ### Changes
 * removed official support for Python 3.7-3.9 (EOL), added Python 3.13 and 3.14
 
+### New features
+* added option `--order-after-ff`, that allows to run `pytest-order` after built-in hooks
+  like the `--failed-first` option  (see [#234](https://github.com/pytest-dev/pytest-order/issues/234))
+
 ### Infrastructure
 * use trusted publisher for release (see https://docs.pypi.org/trusted-publishers/)
 * use `pyproject.toml` for project setup
 
 ### Documentation
 * use a theme for documentation supporting dark mode
-
-### Documentation
 * added use case for ordering test modules (see [#51](https://github.com/pytest-dev/pytest-order/issues/51))
+* fixed documentation for `--indulgent-ordering` option
 
 ## [Version 1.3.0](https://pypi.org/project/pytest-order/1.3.0/) (2024-08-22)
 Allows to fail tests that cannot be ordered.